Backend Operations

Backends in PicoQuant aim to provide the same functionality but each is tailored for a specific purpose/hardware platform/use case. Each backend should implement the same interface. Here you will find documentation on the backend structures and the functions they implement.

Title

PicoQuant.AbstractBackend — Type

Abstract type defining backends which determine the behavior of functions that act on TensorNetworkCircuits

PicoQuant.DSLBackend — Type

The dsl backend will get functions that act on a tensor network circuit to write dsl commands. It also holds filenames for files containing dsl commands, tensor data and the result of fully contracting a tensor network circuit.

PicoQuant.InteractiveBackend — Type

The Interactive backend gets functions to act on a TensorNetworkCircuit interactively.

PicoQuant.save_tensor_data — Function

function save_tensor_data(backend::InteractiveBackend,
                          tensor_label::Symbol,
                          tensor_data::Array{<:Number})

Function to save tensor data to a dictionary in an interactive backend.

function save_tensor_data(backend::DSLBackend,
                          node_label::Symbol,
                          tensor_label::Symbol,
                          tensor_data::Array{<:Number})

Function to save tensor data to the hdf5 file specified by a dsl backend. It also appends a load command to the dsl script contained in the backend to load the saved tensor data as a tensor with the given node label.

function save_tensor_data(network::TensorNetworkCircuit,
                          tensor_label::Symbol,
                          tensor_data::Array{<:Number})

Function to save tensor data to a dictionary in the network's backend.

PicoQuant.load_tensor_data — Function

function load_tensor_data(backend::InteractiveBackend,
                          tensor_label::Symbol)

Function to load tensor data from backend storage (if present)

function load_tensor_data(backend::DSLBackend,
                          tensor_label::Symbol)

Function to load tensor data from backend storage (if present)

function load_tensor_data(network::TensorNetworkCircuit,
                          tensor_label::Symbol)

Function to load tensor data from backend storage (if present)

PicoQuant.save_output — Function

function save_output(backend::InteractiveBackend, node::Symbol, name::String)

Function to save the result of fully contracting a network under the given name.

function save_output(backend::DSLBackend, node::Symbol, name::String)

Function to save the result of fully contracting a network under the given name.

function save_output(network::TensorNetworkCircuit, node::Symbol, name::String)

Function to save the result of fully contracting a network under the given name.

PicoQuant.reshape_tensor — Function

function reshape_tensor(backend::InteractiveBackend, tensor::Symbol, shape)

Function to reshape a given tensor.

function reshape_tensor(backend::DSLBackend, tensor::Symbol, shape)

Function to add dsl command that reshapes a given tensor.

function reshape_tensor(network::TensorNetworkCircuit, tensor::Symbol, shape)

Function to reshape a given tensor.

function reshape_tensor(tensor, dims)

Reshape a tensor

PicoQuant.decompose_tensor! — Function

function decompose_tensor!(backend::InteractiveBackend,
                           tensor::Symbol,
                           left_positions::Array{Int, 1},
                           right_positions::Array{Int, 1};
                           threshold::AbstractFloat=1e-13,
                           max_rank::Integer=0,
                           left_label::Symbol,
                           right_label::Symbol)

Function to decompose a single tensor into two tensors

function decompose_tensor!(backend::DSLBackend,
                           tensor::Symbol,
                           left_indices::Array{Int, 1},
                           right_indices::Array{Int, 1};
                           threshold::AbstractFloat=1e-13,
                           max_rank::Integer=0,
                           left_label::Symbol,
                           right_label::Symbol)

Function to decompose a single tensor into two tensors

function decompose_tensor!(network::TensorNetworkCircuit
                           tensor::Symbol,
                           left_positions::Array{Int, 1},
                           right_positions::Array{Int, 1};
                           threshold::AbstractFloat=1e-13,
                           max_rank::Integer=0,
                           left_label::Symbol,
                           right_label::Symbol)

Function to decompose a single tensor into two tensors

function decompose_tensor!(tng::TensorNetworkCircuit,
                           node::Symbol
                           left_indices::Array{Symbol, 1},
                           right_indices::Array{Symbol, 1};
                           threshold::AbstractFloat=1e-15,
                           max_rank::Integer=0,
                           left_label::Union{Nothing, Symbol}=nothing,
                           right_label::Union{Nothing, Symbol}=nothing)

Decompose a tensor into two smaller tensors

Base.push! — Function

function push!(dsl::DSLBackend, instruction::String)

Function to append a dsl command to the dsl script contained in a dsl backend.