Layer 1 Operations

Layer 1 operations perform the actual computations that are coordinated by layer 2. This layer consists of a number of backends which are capable of performing the computations for a variety of different platforms and scenarios.

Title

PicoQuant.load_tensor! — Function

function load_tensor!(tensors::Dict{Symbol, Array{<:Number}},
                      tensor_label::String,
                      data_label::String,
                      tensor_data_filename::String)

Load tensor data, identified by tensor_label, from a .h5 file and store it in the dictionary 'tensors'

PicoQuant.save_tensor! — Function

function save_tensor!(tensor_data_filename::String,
                      tensors::Dict{Symbol, Array{<:Number}},
                      tensor_label::String,
                      group_name::String)

Save tensor data to a .h5 file. If no group name is given for the data the label that identifies the tensor in the dictionary 'tensors' is used.

PicoQuant.delete_tensor! — Function

function delete_tensor!(backend::InteractiveBackend, tensor_label::Symbol)

Mark tensor for deletion

function delete_tensor!(backend::DSLBackend, tensor_label::Symbol)

Mark tensor for deletion

function delete_tensor!(tensors::Dict{Symbol, Array{<:Number}},
                        tensor_label::String)

Delete the specified tensor from the dictionary 'tensors'

PicoQuant.contract_tensors — Function

function contract_tensors(backend::InteractiveBackend,
                          A_label::Symbol, A_ncon_indices::Array{<:Integer, 1},
                          B_label::Symbol, B_ncon_indices::Array{<:Integer, 1},
                          C_label::Symbol)

Function to interactively contract two tensors A and B to create a tensor C.

function contract_tensors(backend::DSLBackend,
                          A_label::Symbol, A_ncon_indices::Array{<:Integer, 1},
                          B_label::Symbol, B_ncon_indices::Array{<:Integer, 1},
                          C_label::Symbol)

Function to add dsl commands to a dsl backend's script that contract two tensors A and B to create a third tensor C.

function contract_tensors(tensors_to_contract::Tuple{Array{<:Number}, Array{<:Number}},
                          tensor_indices::Tuple{Array{<:Integer,1},Array{<:Integer,1}})

Function to contract the tensors contained in the tuple 'tensorstocontract' according to the ncon indices given and return the result.

Missing docstring.

Missing docstring for PicoQuant.reshape_tensor. Check Documenter's build log for details.

PicoQuant.transpose_tensor — Function

function transpose_tensor(tensor::Array{<:Number},
                          index_permutation::Array{<:Integer, 1})

Transpose a tensor by permuting the indices as specified in index_permutation.

PicoQuant.conjugate_tensor — Function

function conjugate_tensor(tensor::Array{<:Number})

Conjugate the elements of a tensor.

PicoQuant.execute_dsl_file — Function

function execute_dsl_file(dsl_filename::String,
                          tensor_data_filename::String)

Function to read dsl commands from a given dsl file and execute them using tensor data from the given tensor data file.