Layer 2 Operations

Layer 2 operations are concerned with manipulating tensor network structures and are responsible for coordinating and passing off to layer 1 functions to perform the computations.

Title

PicoQuant.random_contraction_plan — Function

function random_contraction_plan(network::TensorNetworkCircuit)

Function to create a random contraction plan

PicoQuant.inorder_contraction_plan — Function

function inorder_contraction_plan(network::TensorNetworkCircuit)

Function to return contraction plan by order of indices

PicoQuant.contraction_plan_to_json — Function

function contraction_plan_to_json(plan::Array)

Function to serialise the contraction plan to json format

PicoQuant.contraction_plan_from_json — Function

function contraction_plan_from_json(str::String)

Function to deserialize the contraction plan from a json string

PicoQuant.contract_pair! — Function

function contract_pair!(network::TensorNetworkCircuit,
                        edge::Symbol)

Contract a pair of nodes connected by the given edge.

function contract_pair!(network::TensorNetworkCircuit,
                        A_label::Symbol,
                        B_label::Symbol)

Function to contract the nodes A and B of the network.

PicoQuant.contract_network! — Function

function contract_network!(network::TensorNetworkCircuit,
                           plan::Array{Symbol, 1},
                           output_shape::Union{String, Array{<:Integer, 1}})

Function to contract the given network according to the given contraction plan. The resulting tensor will be given the shape described by 'output_shape'.

PicoQuant.full_wavefunction_contraction! — Function

function full_wavefunction_contraction!(tn::TensorNetworkCircuit,
                                        output_shape::Union{String, Array{<:Integer, 1}}="")

Function to contract a network by first contracting input nodes together, to get the wavefunction representing the initial state, and then contracting gates into the wavefunction in the order they appear in the circuit.

PicoQuant.compress_tensor_chain! — Function

function compress_tensor_chain!(network::TensorNetworkCircuit,
                                nodes::Array{Symbol, 1};
                                threshold::AbstractFloat=1e-13)

Compress a chain of tensors by given by the array of symbols. This is achieved by peforming forward and backward sweeps of compressing each bond.

PicoQuant.contract_mps_tensor_network_circuit! — Function

function contract_mps_tensor_network_circuit(network::TensorNetworkCircuit;
                                             threshold::AbstractFloat=1e-13)

Contract a tensor network representing a quantum circuit using MPS techniques

PicoQuant.calculate_mps_amplitudes! — Function

function calculate_mps_amplitudes!(network::TensorNetworkCircuit,
                                  mps_nodes::Array{Symbol, 1},
                                  result::String="result")

Calculate amplitudes from an MPS state

PicoQuant.sort_indices — Function

function sort_indices(A::Node, B::Node)

This function divides all of the indices of two nodes A and B into two arrays, one array for all the shared indices between A and B and a second array for all the indices that are not shared.

PicoQuant.create_ncon_indices — Function

function create_ncon_indices(A::Node, B::Node,
                             common_indices::Array{Symbol, 1},
                             uncommon_indices::Array{Symbol, 1})

This function returns ncon indices for the nodes A and B. These can then to be passed to ncon in order to contract the tensor data associated with A and B.