Layer 3 Operations

Layer 3 is the highest layer and deals with the conversion of circuit descriptions (QASM/qiskit circuit objects) to tensor networks. It provides data structures and functionality to represent and manipulate tensor networks representing quantum circuits.

Tensor network fundamentals

PicoQuant.TensorNetworkCircuit — Type

Struct for tensor network graph of a circuit

PicoQuant.Node — Type

Struct to represent a node in tensor network graph

PicoQuant.Edge — Type

Struct to represent an edge

PicoQuant.add_gate! — Function

function add_gate!(network::TensorNetworkCircuit,
                   gate_data::Array{<:Number},
                   targetqubits::Array{Integer, 1};
                   decompose::Bool=false)

Add a node to the tensor network for the given gate acting on the given quibits

PicoQuant.edges — Function

function edges(network::TensorNetworkCircuit)

Function to return edges from a given network

PicoQuant.new_label! — Function

function new_label!(network::TensorNetworkCircuit, label_str)

Function to create a unique symbol by incrememting relevant counter

PicoQuant.add_input! — Function

function add_input!(network::TensorNetworkCircuit, config::String)

Function to add input nodes to a tensor network circuit with a given input configuration

PicoQuant.add_output! — Function

function add_output!(network::TensorNetworkCircuit, config::String)

Function to add output nodes to a tensor network circuit with a given output configuration

PicoQuant.inneighbours — Function

function inneighbours(network::TensorNetworkCircuit,
                      node_label::Symbol)

Function to return the nodes which are connected to the given node with incoming edges

PicoQuant.outneighbours — Function

function outneighbours(network::TensorNetworkCircuit,
                       node_label::Symbol)

Function to return the nodes which are connected to the given node with outgoing edges

PicoQuant.virtualneighbours — Function

function virtualneighbours(network::TensorNetworkCircuit,
                       node_label::Symbol)

Function to return the nodes which are connected to the given node with virtual edges

PicoQuant.neighbours — Function

function neighbours(network::TensorNetworkCircuit,

Function to get all neighbouring nodes of the given node (incoming, outgoing and virtual)

PicoQuant.inedges — Function

function inedges(network::TensorNetworkCircuit,
                 node_label::Symbol)

Function to get all incoming edges to the current node

PicoQuant.outedges — Function

function outedges(network::TensorNetworkCircuit,
                  node_label::Symbol)

Function to get all outgoing edges to the current node

PicoQuant.decompose_gate! — Function

function decompose_gate!(gate_data::Array{<:Number, 4},
                         threshold::AbstractFloat=1e-15)

Decompose a tensor into two smaller tensors

Qiskit/external package interoperability

PicoQuant.load_qasm_as_circuit_from_file — Function

function load_qasm_as_circuit_from_file(qasm_path::String)

Function to load qasm from the given path and return a qiskit circuit

PicoQuant.load_qasm_as_circuit — Function

function load_qasm_as_circuit(qasm_str::String)

Function to load qasm from a given qasm string and return a qiskit circuit

PicoQuant.convert_qiskit_circ_to_network — Function

function convert_qiskit_circ_to_network(circ, backend::AbstractBackend=InteractiveBackend();
                                        decompose::Bool=false,
                                        transpile::Bool=false)

Given a qiskit circuit object, this function will convert this to a tensor network circuit with a backend set to backend. If the decompose option is true it will decompose two qubit gates to two tensors acting on each qubit with a virtual bond connecting them. The transpile option will transpile the circuit using the basicswap pass from qiskit to ensure that two qubit gates are only applied between neighbouring qubits.

Circuit I/O transformers

PicoQuant.to_dict — Function

function to_dict(edge::Edge)

Function to convert an edge instance to a serialisable dictionary

function to_dict(node::Node)

Function to serialise node instance to json format

function to_dict(network::TensorNetworkCircuit)

Convert a tensor network to a nested dictionary

PicoQuant.to_json — Function

function to_json(tng::TensorNetworkCircuit)

Convert a tensor network to a json string

PicoQuant.network_from_dict — Function

function network_from_dict(dict::Dict{String, Any}, backend::AbstractBackend=InteractiveBackend())

Convert a dictionary to a tensor network with a given backend (default: InteractiveBackend)

PicoQuant.edge_from_dict — Function

function edge_from_dict(dict::Dict)

Function to create an edge instance from a dictionary

PicoQuant.node_from_dict — Function

function node_from_dict(d::Dict)

Function to create a node instance from a json string

PicoQuant.network_from_json — Function

function network_from_json(json_str::String, backend::AbstractBackend=InteractiveBackend())

Convert a json string to a tensor network