PicoQuant Visualisation

It can be useful to create visualisations of the tensor network circuit to get insights into the structure and find issues with their manipulation. Some useful functions for this are listed here.

PicoQuant.create_graph — Function

function create_graph(tng::TensorNetworkCircuit)

Create a light graph representation of the tensor network circuit

PicoQuant.plot — Function

function plot(tng::TensorNetworkCircuit; showlabels::Bool=false)

Function for creating a graph of a tensor network graph